@@ -36,7 +36,7 @@ class PersonLayerWritables < PersonFetchables
         joins(roles: :group).
         where(roles: { deleted_at: nil }, groups: { deleted_at: nil }).
         where(conditions.to_a).
-        uniq
+        distinct
     else
       Person.none
     end
