@@ -22,13 +22,13 @@ describe Event::ListsController do
     end
 
     it 'shows error if date_from newer than date to' do
-      get :bsv_export, bsv_export: { date_from: '11.11.2015', date_to: '10.10.2015' }, year: 2016
+      get :bsv_export, params: { bsv_export: { date_from: '11.11.2015', date_to: '10.10.2015' }, year: 2016 }
       is_expected.to redirect_to(list_courses_path)
       expect(flash[:alert]).to eq 'Abschlussdatum von kann nicht neuer als Abschlussdatum bis sein.'
     end
 
     it 'shows error if date_from is invalid' do
-      get :bsv_export, bsv_export: { date_from: '31.06.2015', date_to: '10.10.2015' }, year: 2016
+      get :bsv_export, params: { bsv_export: { date_from: '31.06.2015', date_to: '10.10.2015' }, year: 2016 }
       is_expected.to redirect_to(list_courses_path)
       expect(flash[:alert]).to eq 'Ungültiges Abschlussdatum (von/bis).'
     end
@@ -37,8 +37,8 @@ describe Event::ListsController do
       create_course('123', '03.01.2015', '09.01.2015')
       create_course('124', '11.11.2015', '12.11.2015')
 
-      get :bsv_export, bsv_export: { event_kinds: [kind.id], date_from: '09.09.2015' }, year: 2016
-      expect(response).to be_success
+      get :bsv_export, params: { bsv_export: { event_kinds: [kind.id], date_from: '09.09.2015' }, year: 2016 }
+      expect(response).to be_successful
       expect(rows.size).to eq(2)
       expect(rows.first).to match(/^Vereinbarung-ID-FiVer;Kurs-ID-FiVer;Kursnummer/)
       expect(rows.second).to match(/^fiver42;;124;11.11.2015;/)
@@ -48,8 +48,8 @@ describe Event::ListsController do
       create_course('123', '03.01.2015', '09.01.2015')
       create_course('124', '11.11.2015', '12.11.2015')
 
-      get :bsv_export, bsv_export: { event_kinds: [kind.id], date_to: '09.09.2015' }, year: 2016
-      expect(response).to be_success
+      get :bsv_export, params: { bsv_export: { event_kinds: [kind.id], date_to: '09.09.2015' }, year: 2016 }
+      expect(response).to be_successful
       expect(rows.size).to eq(2)
       expect(rows.second).to match(/^fiver42;;123;03.01.2015;/)
     end
@@ -60,8 +60,8 @@ describe Event::ListsController do
       create_course('125', '06.06.2015', nil)
       create_course('126', '06.06.2015', '08.06.2015', 'created')
 
-      get :bsv_export, bsv_export: { date_from: '01.02.2015', date_to: '21.12.2015' }, year: 2016
-      expect(response).to be_success
+      get :bsv_export, params: { bsv_export: { date_from: '01.02.2015', date_to: '21.12.2015' }, year: 2016 }
+      expect(response).to be_successful
       expect(rows.size).to eq(3)
       expect(rows.second).to match(/^fiver42;;125;06.06.2015;/)
       expect(rows.third).to match(/^fiver42;;124;11.11.2015;/)
@@ -71,8 +71,8 @@ describe Event::ListsController do
       create_course('123', '03.01.2015', '09.01.2015')
       create_course('124', '03.01.2015', '09.01.2015', 'closed', event_kinds(:glk))
 
-      get :bsv_export, bsv_export: { event_kinds: [event_kinds(:glk).id], date_to: '09.09.2015' }, year: 2016
-      expect(response).to be_success
+      get :bsv_export, params: { bsv_export: { event_kinds: [event_kinds(:glk).id], date_to: '09.09.2015' }, year: 2016 }
+      expect(response).to be_successful
       expect(rows.size).to eq(2)
       expect(rows.second).to match(/^fiver42;;124;03.01.2015;/)
     end
@@ -81,8 +81,8 @@ describe Event::ListsController do
       create_course('123', '03.01.2015', '09.01.2015')
       create_course('124', '03.01.2015', '09.01.2015', 'closed', kind, groups(:top_layer))
 
-      get :bsv_export, year: '2015', group_id: groups(:top_layer).id, bsv_export: { event_kinds: [kind.id], date_to: '09.09.2015' }
-      expect(response).to be_success
+      get :bsv_export, params: { year: '2015', group_id: groups(:top_layer).id, bsv_export: { event_kinds: [kind.id], date_to: '09.09.2015' } }
+      expect(response).to be_successful
       expect(rows.size).to eq(2)
       expect(rows.second).to match(/^fiver42;;124;03.01.2015;/)
     end
